Added stores in cartogram
Summary:
Trying to get to a state where we can just require Cartogram and
create objects without depending on a single paper object.

Created stores as singletons that will be accessible by all cartogram
objects. This allows us to avoid crawling upward through a sea of
options and depending on a single renderer instance. On the negative
not all options map to this approach. This works for caches pretty
well. If we need a camera instance or need to know if we are rendering
for webgl or canvas we run into problems. I think there are ways to
mitigate those concerns eventually though.

Test Plan: - Rendered a map in venue_map_renderer

Reviewers: alex, mtoniut

Subscribers: john

Differential Revision: https://phabricator.evbhome.com/D2712