Added interaction priorities to shapes
Sometimes you want shapes, such as text, to appear on top of another, like a circle, but don’t want the circle to lose mouse input. Adding an interaction priority to the circle will ensure when multiple objects come back from the rtree, the circle is sorted ahead of a text label.